Don't generate dummy image during note drawing
Old way: make a dummy request to the normal-sized image so we could calculate the pixel grid for note layout. Didn't generate a duplicate HTTP request because we knew we'd be using the normal size image. Now that we have srcset, the large image might be *actually* used by the browser, so this dummy request can be a duplicate.

New way: use the actual page image instead of a dummy. It means notes don't get drawn until the page image has loaded, but that's fine, they aren't usable until then anyway. Also means that when the large image is fetched, our calculated height might not match the actual grid the notes were laid out on. Still, worth while.

If document JSON returns aspect ratio, we can stop doing all of this.

Affects #50